worked on get, post, put, and delete endpoints for the user model.

+++ shortstories/Controllers/API/UserModelsController.cs
@@ -22,15 +22,16 @@ namespace shortstories.Controllers.API
         }
 
         // GET: api/UserModels
-        [HttpGet]
+        /*[HttpGet]
         public async Task<ActionResult<IEnumerable<UserModel>>> GetUser()
         {
             return await _context.User.ToListAsync();
         }
+        */
 
-        // GET: api/UserModels/5
+        // GET: api/UserModels/{id}
         [HttpGet("{id}")]
-        public async Task<ActionResult<UserModel>> GetUserModel(string id)
+        public async Task<ActionResult<UserModel>> RetrieveAUser(string id)
         {
             var userModel = await _context.User.FindAsync(id);
 
@@ -42,13 +43,59 @@ namespace shortstories.Controllers.API
             return userModel;
         }
 
-        // PUT: api/UserModels/5
+        // PUT: api/UserModels/{id}
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see https://aka.ms/RazorPagesCRUD.
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutUserModel(string id, UserModel userModel)
+        [HttpPut("update/username/{id}")]
+        public async Task<IActionResult> UpdateUserUsername(string id, [FromBody] System.Text.Json.JsonElement jsonData)
         {
-            if (id != userModel.UserModelId)
+            string jsonDataUsername = jsonData.GetProperty("userUsername").ToString();
+
+            var userModel = await _context.User.FindAsync(id);
+            userModel.UserUsername = jsonDataUsername;
+
+            if (userModel == null)
+            {
+                return NotFound();
+            } else if (userModel.UserModelId != id)
+            {
+                return BadRequest();
+            }
+
+            _context.Entry(userModel).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!UserModelExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+            return NoContent();
+        }
+
+        [HttpPut("update/password/{id}")]
+        public async Task<IActionResult> UpdateUserPassword(string id, [FromBody] System.Text.Json.JsonElement jsonData)
+        {
+            string jsonDataPassword = jsonData.GetProperty("userPassword").ToString();
+
+            var userModel = await _context.User.FindAsync(id);
+            userModel.UserPassword = jsonDataPassword;
+
+            if (userModel == null)
+            {
+                return NotFound();
+            }
+            else if (userModel.UserModelId != id)
             {
                 return BadRequest();
             }
@@ -77,16 +124,22 @@ namespace shortstories.Controllers.API
         // POST: api/UserModels
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see https://aka.ms/RazorPagesCRUD.
-        [HttpPost]
-        public async Task<ActionResult<UserModel>> PostUserModel(UserModel userModel)
+        [HttpPost("register")]
+        public async Task<ActionResult<UserModel>> RegisterAUser([FromBody] System.Text.Json.JsonElement jsonData)
         {
+            string jsonDataUsername = jsonData.GetProperty("userUsername").ToString();
+            string jsonDataPassword = jsonData.GetProperty("userPassword").ToString();
+
+            UserModel userModel = new UserModel(jsonDataUsername, jsonDataPassword);
             _context.User.Add(userModel);
+
             try
             {
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateException)
             {
+                
                 if (UserModelExists(userModel.UserModelId))
                 {
                     return Conflict();
@@ -100,18 +153,46 @@ namespace shortstories.Controllers.API
             return CreatedAtAction("GetUserModel", new { id = userModel.UserModelId }, userModel);
         }
 
-        // DELETE: api/UserModels/5
-        [HttpDelete("{id}")]
-        public async Task<ActionResult<UserModel>> DeleteUserModel(string id)
+        [HttpPost("login")]
+        public async Task<ActionResult<UserModel>> LoginAUser([FromBody] System.Text.Json.JsonElement jsonData)
         {
-            var userModel = await _context.User.FindAsync(id);
+            string jsonDataUsername = jsonData.GetProperty("userUsername").ToString();
+            string jsonDataPassword = jsonData.GetProperty("userPassword").ToString();
+
+            var userModel = await _context.User.Where(o => o.UserUsername == jsonDataUsername).Where(p => p.UserPassword == jsonDataPassword).FirstOrDefaultAsync();
+
+            if (userModel == null)
+            {
+                return BadRequest();
+            }
+
+            return userModel;
+        }
+
+        // DELETE: api/UserModels/{id}
+        [HttpDelete("{id}")]
+        public async Task<ActionResult<UserModel>> DeleteAUser(string id, [FromBody] System.Text.Json.JsonElement jsonData)
+        {
+            if (jsonData.Equals(null))
+            {
+                return NotFound();
+            }
+
+            string password = jsonData.GetProperty("userPassword").ToString();
+
+            var userModel = await _context.User.Where(o => o.UserPassword == password).Where(p => p.UserModelId == id).FirstOrDefaultAsync();
+            System.Diagnostics.Debug.WriteLine("LOOK UNDER");
+            System.Diagnostics.Debug.WriteLine(userModel);
             if (userModel == null)
             {
                 return NotFound();
             }
 
-            _context.User.Remove(userModel);
-            await _context.SaveChangesAsync();
+            if (userModel.UserPassword == password)
+            {
+                _context.User.Remove(userModel);
+                await _context.SaveChangesAsync();
+            }
 
             return userModel;
         }
